clients, simulators: util refactor and stability (#1458)
* refactor + stability

* lint

* helper runtime
