+++ clients/nlean/nlean.sh
@@ -7,6 +7,31 @@ NODE_ID="${HIVE_NODE_ID:-nlean_0}"
 ASSET_ROOT="/tmp/nlean-runtime"
 FORK_DIGEST="${HIVE_LEAN_FORK_DIGEST:-12345678}"
 KEY_FILE=""
+LOCAL_IP_PLACEHOLDER="__HIVE_LOCAL_IP__"
+
+detect_local_ip() {
+    hostname -i 2>/dev/null | tr ' ' '\n' | grep -E '^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$' | grep -v '^127\.' | head -n 1
+}
+
+materialize_runtime_local_ip() {
+    local runtime_ip
+
+    if [ ! -f "$ASSET_ROOT/validator-config.yaml" ]; then
+        return
+    fi
+
+    if ! grep -q "$LOCAL_IP_PLACEHOLDER" "$ASSET_ROOT/validator-config.yaml"; then
+        return
+    fi
+
+    runtime_ip="$(detect_local_ip || true)"
+    if [ -z "$runtime_ip" ]; then
+        echo "Unable to resolve local container IP for $NODE_ID" >&2
+        exit 1
+    fi
+
+    sed -i "s/${LOCAL_IP_PLACEHOLDER}/${runtime_ip}/g" "$ASSET_ROOT/validator-config.yaml"
+}
 
 case "$DEVNET_LABEL" in
     devnet4)
@@ -43,10 +68,17 @@ if [ ! -f "$ASSET_ROOT/config.yaml" ]; then
     exit 1
 fi
 
+if [ ! -f "$ASSET_ROOT/nodes.yaml" ]; then
+    echo "Missing prepared Lean bootnodes config at $ASSET_ROOT/nodes.yaml" >&2
+    exit 1
+fi
+
 if [ ! -d "$ASSET_ROOT/hash-sig-keys" ]; then
     echo "Missing hash-sig-keys directory at $ASSET_ROOT/hash-sig-keys" >&2
     exit 1
 fi
+
+materialize_runtime_local_ip
 
 # nlean reads the libp2p identity via --node-key pointing at a raw hex private
 # key file. Prefer HIVE_CLIENT_PRIVATE_KEY if the sim supplies one, otherwise
@@ -65,12 +97,10 @@ else
 fi
 
 FLAGS=(
-    --validator-config "$ASSET_ROOT/validator-config.yaml"
-    --annotated-validators "$ASSET_ROOT/annotated_validators.yaml"
-    --hash-sig-key-dir "$ASSET_ROOT/hash-sig-keys"
+    --custom-network-config-dir "$ASSET_ROOT"
     --node "$NODE_ID"
     --data-dir /data
-    --fork-digest "$FORK_DIGEST"
+    --network "$FORK_DIGEST"
     --node-key "$NODE_KEY_PATH"
     --socket-port 9000
     --metrics-address 0.0.0.0
