+++ simulators/lean/src/scenarios/mod.rs
@@ -1,6 +1,5 @@
 pub mod client_interop;
 pub(crate) mod gossip;
-pub(crate) mod helper;
 pub(crate) mod reqresp;
 pub mod rpc_compat;
 pub mod sync;
