+++ simulators/lean/src/utils/mod.rs
@@ -1,1 +1,3 @@
+pub(crate) mod helper;
 pub mod libp2p_mock;
+pub(crate) mod util;
